SLVS-1816 Move MuteIssueService and its models in ConnectedMode assembly
This fixes architectural violation of Core having a refernce to SonarQube.Client
