+++ src/Integration.UnitTests/Transition/MuteIssuesWindowServiceTests.cs
@@ -18,9 +18,9 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
+using SonarLint.VisualStudio.ConnectedMode.Transition;
 using SonarLint.VisualStudio.Core;
 using SonarLint.VisualStudio.Core.Binding;
-using SonarLint.VisualStudio.Core.Transition;
 using SonarLint.VisualStudio.Integration.Transition;
 using SonarLint.VisualStudio.TestInfrastructure;
